ServerManager compatibility fixes.

+++ Patches/TrackingPatches.cs
@@ -38,8 +38,10 @@ namespace RtDQuestForge.Patches
             {
                 // A remote player's kill died on our machine (or on the
                 // dedicated server). Route the credit to the killer.
-                // DEBUG breadcrumb, remove once kill routing is verified.
-                Jotunn.Logger.LogMessage("Routing kill credit: " + attackerPlayer.GetPlayerName() + " killed " + prefabName);
+                if (QuestForgePlugin.VerboseLogging)
+                {
+                    Jotunn.Logger.LogMessage("Routing kill credit: " + attackerPlayer.GetPlayerName() + " killed " + prefabName);
+                }
 
                 QuestSync.SendKillCredit(attackerPlayer.GetPlayerName(), prefabName);
             }
